+++ src/test/kotlin/tornadofx/testapps/FormTest.kt
@@ -7,18 +7,18 @@ class FormApp : App(FormView::class)
 
 class FormView : View("My Form") {
     override val root = form {
-        fieldset("FieldSet") {
-            labelPosition = Orientation.VERTICAL
+        fieldset("FieldSet", labelPosition = Orientation.VERTICAL) {
+            wrapWidth = 340
 
-            field("_Field 1") {
-                textarea() {
+            field("Field 1") {
+                textarea {
                     prefRowCount = 2
                     mnemonicTarget()
                 }
             }
 
             field("Field 2") {
-                textarea() {
+                textarea {
                     prefRowCount = 10
                 }
             }