+++ CHANGELOG.md
@@ -56,6 +56,7 @@ All notable changes to this project will be documented in this file.
 - `cellCache` supports builders. Earlier, builders would be appended to the ListView, creating undesirable results
 - `Scene.reloadViews()` is removed from the public API, no need to call it manually
 - `titledpane` builder now accepts op parameter like every other builder. node parameter is now optional
+- Fieldset.wrapWidth is now Number instead of Double
 
 ## [1.5.9]